# docs/swagger.yaml
definitions:
  model.DeviceInfo:
    properties:
      balance:
        type: string
      baud_rate:
        type: integer
      connected:
        type: boolean
      error:
        type: string
      imei:
        type: string
      imsi:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      network_type:
        type: string
      operator:
        type: string
      phone_number:
        type: string
      port:
        type: string
      signal_level:
        type: integer
      timestamp:
        type: string
      version:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      timestamp:
        type: string
    type: object
  model.HealthResponse:
    properties:
      status:
        type: string
      timestamp:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
  model.ModemInfo:
    properties:
      baud_rate:
        type: integer
      connected:
        type: boolean
      imei:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      port:
        type: string
      signal:
        type: integer
      version:
        type: string
    type: object
  model.PortStatus:
    properties:
      available:
        type: boolean
      balance:
        type: string
      error:
        type: string
      in_use:
        type: boolean
      port:
        type: string
    type: object
  model.SendSMSRequest:
    properties:
      baud_rate:
        type: integer
      message:
        type: string
      mode:
        description: '"text" or "pdu", default "text"'
        type: string
      port:
        type: string
      priority:
        description: '"normal", "high", "urgent"'
        type: string
      timeout:
        type: integer
      to:
        type: string
    required:
    - message
    - to
    type: object
  model.SendSMSResponse:
    properties:
      duration:
        type: string
      error:
        type: string
      message:
        type: string
      message_id:
        type: string
      mode:
        description: '"text" or "pdu"'
        type: string
      steps:
        items:
          type: string
        type: array
      success:
        type: boolean
      timestamp:
        type: string
      to:
        type: string
    type: object
  model.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    get:
      description: Get information about the SMS Gateway API
      produces:
      - application/json
      responses:
        "200":
          description: API information
          schema:
            additionalProperties: true
            type: object
      summary: API information
      tags:
      - General
  /api/v1/device/info:
    get:
      description: Get comprehensive device information including phone number, balance,
        network type, and SIM details
      parameters:
      - description: Port name (defaults to configured default port)
        in: query
        name: port
        type: string
      - description: Baud rate (defaults to configured default baud rate)
        in: query
        name: baud_rate
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detailed device information
          schema:
            $ref: '#/definitions/model.DeviceInfo'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get detailed device information
      tags:
      - Device
  /api/v1/health:
    get:
      description: Check the health status of the SMS Gateway service
      produces:
      - application/json
      responses:
        "200":
          description: Service health status
          schema:
            $ref: '#/definitions/model.HealthResponse'
      summary: Health check
      tags:
      - Health
  /api/v1/modem/info:
    get:
      description: Get detailed information about the modem
      parameters:
      - description: Port name (defaults to configured default port)
        in: query
        name: port
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Modem information
          schema:
            $ref: '#/definitions/model.ModemInfo'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get modem information
      tags:
      - Modem
  /api/v1/ports:
    get:
      description: Get a list of all available serial ports with device information
      produces:
      - application/json
      responses:
        "200":
          description: List of available ports with device info
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: List available ports
      tags:
      - Modem
  /api/v1/ports/status:
    get:
      description: Check the status of a specific serial port and return SIM balance
        (if configured)
      parameters:
      - description: Port name (defaults to configured default port)
        in: query
        name: port
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Port status information including balance if available
          schema:
            $ref: '#/definitions/model.PortStatus'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Check port status
      tags:
      - Modem
  /api/v1/sms/send:
    post:
      consumes:
      - application/json
      description: Send an SMS message through the configured modem
      parameters:
      - description: SMS request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SendSMSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: SMS sent successfully
          schema:
            $ref: '#/definitions/model.SendSMSResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SendSMSResponse'
      summary: Send SMS message
      tags:
      - SMS
swagger: "2.0"